requirements.txt
matplotlib==3.10.5
pandas==2.3.1
PyYAML==6.0.2
PyYAML==6.0.2
streamlit==1.44.1
